Remove other dependencies (Issue #29)
- controller 객체들의 다른 객체들에 대한 의존성 제거
