+++ src/main/java/com/example/medicalqnaclient/controller/HelloController.java
@@ -1,0 +1,17 @@
+package com.example.medicalqnaclient.controller;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+/**
+ * 프로그램 실행 첫 화면
+ */
+public class HelloController {
+    @FXML
+    private Label welcomeText;
+
+    // 홈 화면 요청
+    @FXML
+    protected void onHelloButtonClick() {
+    }
+}
