+++ src/main/java/com/example/medicalqnaclient/controller/PatientSignUpController.java
@@ -1,0 +1,25 @@
+package com.example.medicalqnaclient.controller;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.TextField;
+
+/**
+ * 환자 회원 가입 화면
+ */
+public class PatientSignUpController {
+    @FXML
+    private TextField id;
+
+    @FXML
+    private TextField password;
+
+    // 홈 화면 요청
+    @FXML
+    protected void onHomeButtonClick() {
+    }
+
+    // 환자 회원 가입 요청
+    @FXML
+    protected void onSignUpButtonClick() {
+    }
+}
