+++ src/main/java/com/example/medicalqnaclient/controller/QuestionController.java
@@ -1,0 +1,40 @@
+package com.example.medicalqnaclient.controller;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+/**
+ * 질문 조회 화면
+ */
+public class QuestionController {
+    @FXML
+    private Label titleLabel;
+
+    @FXML
+    private Label idLabel;
+
+    @FXML
+    private Label symptomLabel;
+
+    @FXML
+    private Label contentLabel;
+
+    @FXML
+    protected void initialize() {
+    }
+
+    // 홈 화면 요청
+    @FXML
+    protected void onHomeButtonClick() {
+    }
+
+    // 질문 수정 요청
+    @FXML
+    protected void onEditButtonClick() {
+    }
+
+    // 질문 삭제 요청
+    @FXML
+    protected void onDeleteButtonClick() {
+    }
+}
