+++ src/main/java/com/example/medicalqnaclient/controller/QuestionListLoginedController.java
@@ -1,0 +1,37 @@
+package com.example.medicalqnaclient.controller;
+
+import com.example.medicalqnaclient.v1.businesslogic.question.QuestionTitle;
+import javafx.fxml.FXML;
+import javafx.scene.control.*;
+
+/**
+ * 로그인 후 질문 목록 조회 화면
+ */
+public class QuestionListLoginedController {
+
+    @FXML
+    private ComboBox<String> categoryComboBox;
+
+    @FXML
+    private ListView<QuestionTitle> listView;
+
+    @FXML
+    private MenuButton userNameMenuButton;
+
+    @FXML
+    private TextArea userQueryTextArea;
+
+    @FXML
+    protected void initialize() {
+    }
+
+    // 검색 요청
+    @FXML
+    protected void onSearchButtonClick() {
+    }
+
+    // 로그아웃 요청
+    @FXML
+    protected void onLogoutButtonClick() {
+    }
+}
