+++ src/main/java/com/example/medicalqnaclient/controller/QuestionPostController.java
@@ -1,0 +1,29 @@
+package com.example.medicalqnaclient.controller;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.*;
+
+/**
+ * 질문 등록 화면
+ */
+public class QuestionPostController {
+
+    @FXML
+    private TextField titleTextField;
+
+    @FXML
+    private TextArea symptomTextArea;
+
+    @FXML
+    private TextArea questionTextArea;
+
+    // 홈 화면 요청
+    @FXML
+    protected void onHomeButtonClick() {
+    }
+
+    // 질문 등록 요청
+    @FXML
+    private void onWriteButtonClick() {
+    }
+}
