+++ src/main/java/com/example/medicalqnaclient/controller/SystemMonitoringController.java
@@ -1,0 +1,14 @@
+package com.example.medicalqnaclient.controller;
+
+import javafx.fxml.FXML;
+
+/**
+ * 운영 첫 화면
+ */
+public class SystemMonitoringController {
+
+    // 홈 화면 요청 
+    @FXML
+    protected void onHomeButtonClick() {
+    }
+}
