--- src/main/java/com/example/medicalqnaclient/v1/businesslogic/board/menu/DoctorMenu.java
@@ -1,14 +1,0 @@
-package com.example.medicalqnaclient.v1.businesslogic.board.menu;
-
-import com.example.medicalqnaclient.v1.businesslogic.board.menu.factorymethod.MenuComponent;
-import javafx.scene.control.MenuButton;
-
-public class DoctorMenu extends MenuComponent {
-    public DoctorMenu(MenuButton userNameMenuButton) {
-        super(userNameMenuButton);
-    }
-
-    @Override
-    protected void setMenuItem() {
-    }
-}
