--- src/main/java/com/example/medicalqnaclient/v1/businesslogic/board/observerpattern/Observer.java
@@ -1,5 +1,0 @@
-package com.example.medicalqnaclient.v1.businesslogic.board.observerpattern;
-
-public interface Observer {
-    void update(Subject subject);
-}
