--- src/main/java/com/example/medicalqnaclient/v1/businesslogic/board/observerpattern/Subject.java
@@ -1,24 +1,0 @@
-package com.example.medicalqnaclient.v1.businesslogic.board.observerpattern;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public abstract class Subject {
-    private List<Observer> observers = new ArrayList<>();
-
-    public void add(Observer observer) {
-        observers.add(observer);
-    }
-
-    public void remove(Observer observer) {
-        observers.remove(observer);
-    }
-
-    public void notifyObservers() {
-        for (Observer observer: observers) {
-            observer.update(this);
-        }
-    }
-
-    public abstract String getState();
-}
