--- src/main/java/com/example/medicalqnaclient/v1/controllers/admin/SystemMonitoringController.java
@@ -1,17 +1,0 @@
-package com.example.medicalqnaclient.v1.controllers.admin;
-
-import com.example.medicalqnaclient.HelloApplication;
-import javafx.fxml.FXML;
-
-import java.io.IOException;
-
-public class SystemMonitoringController {
-    @FXML
-    protected void onHomeButtonClick() {
-        try {
-            HelloApplication.setRoot("question-list-login-view.fxml");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
