+++ src/main/java/module-info.java
@@ -15,18 +15,6 @@ module com.example.medicalqnaclient {
 
     opens com.example.medicalqnaclient to javafx.fxml;
     exports com.example.medicalqnaclient;
-    exports com.example.medicalqnaclient.v1.controllers;
-    opens com.example.medicalqnaclient.v1.controllers to javafx.fxml;
-    exports com.example.medicalqnaclient.v1.controllers.user.login;
-    opens com.example.medicalqnaclient.v1.controllers.user.login to javafx.fxml;
-    exports com.example.medicalqnaclient.v1.controllers.post;
-    opens com.example.medicalqnaclient.v1.controllers.post to javafx.fxml;
-    exports com.example.medicalqnaclient.v1.controllers.admin;
-    opens com.example.medicalqnaclient.v1.controllers.admin to javafx.fxml;
-    exports com.example.medicalqnaclient.v1.controllers.board;
-    opens com.example.medicalqnaclient.v1.controllers.board to javafx.fxml;
-    exports com.example.medicalqnaclient.v1.controllers.user.signup;
-    opens com.example.medicalqnaclient.v1.controllers.user.signup to javafx.fxml;
-    exports com.example.medicalqnaclient.v1.controllers.user.mydata;
-    opens com.example.medicalqnaclient.v1.controllers.user.mydata to javafx.fxml;
+    exports com.example.medicalqnaclient.controller;
+    opens com.example.medicalqnaclient.controller to javafx.fxml;
 }
